💄 Split "create new job" into page and method,
