+++ libs/consumer/lib/pages/job_listings.dart
@@ -1,6 +1,6 @@
 import 'package:async_redux/async_redux.dart';
+import 'package:consumer/methods/populate_job_listings.dart';
 import 'package:flutter/material.dart';
-// import '../methods/populate_job_listings.dart';
 import 'package:redux_comp/redux_comp.dart';
 
 class ConsumerListings extends StatelessWidget {
@@ -11,21 +11,7 @@ class ConsumerListings extends StatelessWidget {
   Widget build(BuildContext context) {
     return StoreProvider<AppState>(
       store: store,
-      child: MaterialApp(
-        home: Scaffold(
-          appBar: AppBar(
-            title: const Text('My Job Listings'),
-            backgroundColor: const Color.fromRGBO(82, 121, 111, 1),
-          ),
-          // body:
-          // StoreConnector<AppState, VoidCallback> (
-          //   converter: (store) => store.state.adverts,
-          //   builder: (context, adverts) {
-          //     return JobListings(store: store, adverts: adverts);
-          //   },
-        ),
-      ),
-      // ),
+      child: MaterialApp(home: JobListings(store: store)),
     );
   }
 }
